Fixing the getById and deleteById methods from IBookingRepository file

+++ src/main/java/dev/waamir/hotelaluraapi/domain/port/IBookingRepository.java
@@ -7,8 +7,8 @@ import java.util.Optional;
 
 public interface IBookingRepository<T extends Booking> {
     T create(T t);
-    Optional<T> getById(int id);
-    void deleteById(int id);
+    Optional<T> getById(Long id);
+    void deleteById(Long id);
     void update(T t);
     List<T> list();
 }
